feat: prevent loss of an unsaved entry.
When creating a new entry, if a user click on another entry to view
it, ask him to confirm his action to prevent a data loss.

Update the selectEntry redux action to check if the user is creating a new
entry.

Add a new argument to this redux action to be sure to be correctly
redirected to the new entry when the user confirm the creation of
a new entry.

Fixes: https://github.com/buttercup/buttercup-desktop/issues/634