feat: select the archive and display her name when opening it
when opening an archive (on the sidebar), pass the item to active to
show to the user the selected item. Pass the archive name to the
password modal to display the archive name.
manage the selected archive from modalRequest state from the workspace and
pass the info to the sidebar to make it active. retrieve and pass the
name to the password modal to be able to display the archive name.

Fixes: https://github.com/buttercup/buttercup-desktop/issues/812