Adding Support for Portuguese Language (pt-br) and Language Choice Option in the Program
In this commit, I have added the Portuguese language (pt-br) to the 'pt-br.json' file to make our program more accessible to Portuguese speakers. Additionally, we have updated the 'index.ts' file to allow users to choose the pt-br language as an language option. This update is an important step in the project accessibility and will help expand the reach of the application to a broader audience.